Move utils::arrow_* functions to test::fixtures
Since these are test only functions, it's better to keep them in a test
specific module.
